fix UT, test=develop
